requirements.txt
django
south
django-selector
django-annoying
django-extensions
ipdb
